+++ packages/maps/src/ht-maps.ts
@@ -19,6 +19,7 @@ export * from "./entities/segment-polylines";
 export * from "./entities/stop-markers";
 export * from "./entities/current-user";
 export * from "./entities/actions-heatmap";
+export * from "./entities/stops-heatmap";
 
 //mixins
 export * from "./mixins/compounds-data-observable";